PldmRebase

+++ libpldmresponder/pdr_state_effecter.hpp
@@ -45,8 +45,7 @@ void generateStateEffecterPDR(const DBusInterface& dBusIntf, const Json& json,
             {
                 error(
                     "Malformed PDR JSON return pdrEntry; no state set info for state effecter pdr '{STATE_EFFECTER_PDR}'",
-                    "STATE_EFFECTER_PDR",
-                    static_cast<unsigned>(PLDM_STATE_EFFECTER_PDR));
+                    "STATE_EFFECTER_PDR", PLDM_STATE_EFFECTER_PDR);
                 throw InternalFailure();
             }
             pdrSize += sizeof(state_effecter_possible_states) -
@@ -113,8 +112,8 @@ void generateStateEffecterPDR(const DBusInterface& dBusIntf, const Json& json,
 
         pldm::responder::pdr_utils::DbusMappings dbusMappings{};
         pldm::responder::pdr_utils::DbusValMaps dbusValMaps{};
-        uint8_t* start = entry.data() + sizeof(pldm_state_effecter_pdr) -
-                         sizeof(uint8_t);
+        uint8_t* start =
+            entry.data() + sizeof(pldm_state_effecter_pdr) - sizeof(uint8_t);
         for (const auto& effecter : effecters)
         {
             auto set = effecter.value("set", empty);
@@ -148,8 +147,8 @@ void generateStateEffecterPDR(const DBusInterface& dBusIntf, const Json& json,
             pldm::utils::DBusMapping dbusMapping{};
             try
             {
-                auto service = dBusIntf.getService(objectPath.c_str(),
-                                                   interface.c_str());
+                auto service =
+                    dBusIntf.getService(objectPath.c_str(), interface.c_str());
 
                 dbusMapping = pldm::utils::DBusMapping{
                     objectPath, interface, propertyName, propertyType};
