+++ oem/ibm/libpldmresponder/file_io_type_pcie.cpp
@@ -1,16 +1,17 @@
 #include "file_io_type_pcie.hpp"
 
-#include "libpldm/base.h"
-
 #include "common/utils.hpp"
 #include "host-bmc/dbus/custom_dbus.hpp"
 #include "host-bmc/dbus/serialize.hpp"
 #include "utils.hpp"
 
+#include <libpldm/base.h>
 #include <libpldm/oem/ibm/file_io.h>
-#include <stdint.h>
+#include <sys/stat.h>
 
 #include <phosphor-logging/lg2.hpp>
+
+#include <cstdint>
 
 PHOSPHOR_LOG2_USING;
 
@@ -174,7 +175,7 @@ int PCIeInfoHandler::write(const char* buffer, uint32_t, uint32_t& length,
     catch (const std::exception& e)
     {
         error("Create/Write data to the File type {TYPE}, failed {ERROR}",
-              "TYPE", (int)infoType, "ERROR", e);
+              "TYPE", infoType, "ERROR", e);
         return PLDM_ERROR;
     }
 
@@ -507,13 +508,11 @@ void PCIeInfoHandler::setTopologyOnSlotAndAdapter(
     }
 }
 
-void PCIeInfoHandler::parsePrimaryLink(uint8_t linkType,
-                                       const IoSlotLocation& ioSlotLocationCode,
-                                       const LocalPort& localPortLocation,
-                                       const uint32_t& linkId,
-                                       const std::string& linkStatus,
-                                       uint8_t curLinkSpeed, int64_t linkWidth,
-                                       uint8_t parentLinkId)
+void PCIeInfoHandler::parsePrimaryLink(
+    uint8_t linkType, const IoSlotLocation& ioSlotLocationCode,
+    const LocalPort& localPortLocation, const uint32_t& linkId,
+    const std::string& linkStatus, uint8_t curLinkSpeed, int64_t linkWidth,
+    uint8_t parentLinkId)
 {
     // Check the io_slot_location_code size
     if (!ioSlotLocationCode.size())
@@ -900,8 +899,8 @@ void PCIeInfoHandler::parseTopologyData()
     };
 
     // memory map the topology file into pldm memory
-    void* fileInMemory = mmap(NULL, sb.st_size, PROT_READ, MAP_PRIVATE,
-                              topologyFd(), 0);
+    void* fileInMemory =
+        mmap(NULL, sb.st_size, PROT_READ, MAP_PRIVATE, topologyFd(), 0);
     if (MAP_FAILED == fileInMemory)
     {
         error("mmap on topology file failed with error {RC}", "RC", -errno);
@@ -1050,9 +1049,9 @@ void PCIeInfoHandler::parseTopologyData()
                 static_cast<int>(slotLocCodeCompartSize));
         std::string slotLocationCode(slotLocation.begin(), slotLocation.end());
 
-        uint8_t* suffixData = reinterpret_cast<uint8_t*>(slotData) +
-                              slotLocationDataMemberSize +
-                              slotData->slotLocCodesCmnPrtSize;
+        uint8_t* suffixData =
+            reinterpret_cast<uint8_t*>(slotData) + slotLocationDataMemberSize +
+            slotData->slotLocCodesCmnPrtSize;
         if (!suffixData)
         {
             error("slot location suffix data is nullptr");
@@ -1086,8 +1085,8 @@ void PCIeInfoHandler::parseTopologyData()
 
                 slotSuffixLocationCode = slotSuffLocationCode;
             }
-            std::string slotFullLocationCode = slotLocationCode +
-                                               slotSuffixLocationCode;
+            std::string slotFullLocationCode =
+                slotLocationCode + slotSuffixLocationCode;
 
             slotFinaLocationCode.push_back(slotFullLocationCode);
 
@@ -1096,14 +1095,14 @@ void PCIeInfoHandler::parseTopologyData()
         }
 
         // store the information into a map
-        topologyInformation[linkId] =
-            std::make_tuple(linkStateMap[linkStatus], type, linkSpeed,
-                            linkWidth[width], pcieHostBridgeLocationCode,
-                            std::make_pair(localTopPortLocationCode,
-                                           localBottomPortLocationCode),
-                            std::make_pair(remoteTopPortLocationCode,
-                                           remoteBottomPortLocationCode),
-                            slotFinaLocationCode, parentLinkId);
+        topologyInformation[linkId] = std::make_tuple(
+            linkStateMap[linkStatus], type, linkSpeed, linkWidth[width],
+            pcieHostBridgeLocationCode,
+            std::make_pair(localTopPortLocationCode,
+                           localBottomPortLocationCode),
+            std::make_pair(remoteTopPortLocationCode,
+                           remoteBottomPortLocationCode),
+            slotFinaLocationCode, parentLinkId);
 
         // move the pointer to next link
         singleEntryData = reinterpret_cast<struct pcieLinkEntry*>(
@@ -1111,7 +1110,7 @@ void PCIeInfoHandler::parseTopologyData()
             htobe16(singleEntryData->entryLength));
     }
     // Need to call cable info at the end , because we dont want to parse
-    // cable info without parsing the successfull topology successfully
+    // cable info without parsing the successful topology successfully
     // Having partial information is of no use.
     parseCableInfo();
 }
@@ -1144,8 +1143,8 @@ void PCIeInfoHandler::parseCableInfo()
         munmap(fileInMemory, sb.st_size);
     };
 
-    void* fileInMemory = mmap(NULL, sb.st_size, PROT_READ, MAP_PRIVATE,
-                              cableInfoFd(), 0);
+    void* fileInMemory =
+        mmap(NULL, sb.st_size, PROT_READ, MAP_PRIVATE, cableInfoFd(), 0);
 
     if (MAP_FAILED == fileInMemory)
     {
@@ -1190,9 +1189,9 @@ void PCIeInfoHandler::parseCableInfo()
                 htobe16(cableData->ioEnclosurePortLocationCodeOffset),
             cableData->ioEnclosurePortLocationCodeSize);
 
-        std::string cablePartNum(cableDataPtr +
-                                     htobe16(cableData->cablePartNumberOffset),
-                                 cableData->cablePartNumberSize);
+        std::string cablePartNum(
+            cableDataPtr + htobe16(cableData->cablePartNumberOffset),
+            cableData->cablePartNumberSize);
 
         // cache the data into a map
         cableInformation[cable] = std::make_tuple(
